doc(readme): prevent infinite loop when request fails
The code example for `Automatic sign-in` was creating an infinite loop triggering hundreds of signin requests per seconds when the network was unreachable.

The reason is that `auth.signingRedirect()` changes the state of `auth.isLoading`; then, when the request fails because the user is offline, `auth.isLoading` changes again, triggering `auth.signinRedirect()`.